Simplicity by Design By Adam Bien
http://www.oracle.com/technetwork/issue-archive/2011/11-jan/o11java-195110.html

To illustrate, consider a simple MessageMe application that stores a message string in a database. The JSF 2 markup consists of two components: inputText and commandButton. As shown in Listing 2, inputText is value-bound to a class with the name index with a property message, which has, in turn, a content attribute. commandButtons’s action attribute is bound to the save method of the backing bean with the name index.